test: refactor by defining fixtures in a conftest file
