+++ tests/conftest.py
@@ -1,0 +1,20 @@
+import anndata as ad
+import pytest
+from pydeseq2.utils import load_example_data
+
+
+@pytest.fixture
+def test_adata():
+    counts = load_example_data(
+        modality="raw_counts",
+        dataset="synthetic",
+        debug=False,
+    )
+
+    metadata = load_example_data(
+        modality="metadata",
+        dataset="synthetic",
+        debug=False,
+    )
+
+    return ad.AnnData(X=counts, obs=metadata)
