+++ tests/test_de.py
@@ -1,9 +1,7 @@
-import anndata as ad
 import numpy as np
 import pytest
 import statsmodels.api as sm
 from pandas import testing as tm
-from pydeseq2.utils import load_example_data
 
 import multi_condition_comparisions
 from multi_condition_comparisions.tl.de import BaseMethod, StatsmodelsDE
@@ -11,23 +9,6 @@ from multi_condition_comparisions.tl.de import BaseMethod, StatsmodelsDE
 
 def test_package_has_version():
     assert multi_condition_comparisions.__version__ is not None
-
-
-@pytest.fixture
-def test_adata():
-    counts = load_example_data(
-        modality="raw_counts",
-        dataset="synthetic",
-        debug=False,
-    )
-
-    metadata = load_example_data(
-        modality="metadata",
-        dataset="synthetic",
-        debug=False,
-    )
-
-    return ad.AnnData(X=counts, obs=metadata)
 
 
 @pytest.mark.parametrize(
